Moved into caliper-connector directory, with thanks to @etphipps
